third_party/WebKit/LayoutTests/fast/css/sticky/parsing-position-sticky-expected.txt

This tests checks that the sticky value for position parses correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("position: sticky;") is "sticky"
PASS successfullyParsed is true

TEST COMPLETE
